requirements.txt
PyQt6
paddleocr
rapid_latex_ocr
numpy
opencv-python
Pillow
keyboard